requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
reportlab==4.2.2
python-docx==1.1.2
PyMuPDF==1.24.9
Pillow==12.1.0
boto3==1.35.99
